+++ pow/CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `NonceProvider, NonceProviderBuilder` traits;
+- `providers` module;
 
 ## 1.0.0-rc.1 - 2022-10-25