+++ pow/CHANGELOG.md
@@ -32,8 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- `NonceProvider` and `NonceProviderBuilder` traits;
 - `providers` module;
+- `NonceProvider` and `NonceProviderBuilder` traits;
+- Deprecated `pow_hash` and `compute_pow_score`;
 
 ## 1.0.0-rc.1 - 2022-10-25