+++ pow/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `SingleThreadedMinerBuilder` and `SingleThreadedMiner` for `target_family = "wasm"`;
 - `get_miner` and `get_miner_num_workers` functions;
+- `MinerBuilder::new` method;
 
 ### Changed